Fix models

+++ mason_templates/digit_entity/hooks/post_gen.dart
@@ -9,18 +9,29 @@ void run(HookContext context) async {
   final variables = context.vars;
 
   ConfigModel model = Mapper.fromMap<ConfigModel>(variables);
-  if (model.createRepository) {
-    return;
+  if (!model.createRepository) {
+    final path = p.join(
+      'data',
+      'repositories',
+      'remote',
+      '${model.name.snakeCase}.dart',
+    );
+    await _deleteFile(path);
   }
 
-  final path = p.join(
-    'data',
-    'repositories',
-    'remote',
-    '${model.name.snakeCase}.dart',
-  );
-  final file = File(path);
-  if (file.existsSync()) {
-    await file.delete();
+  if (model.isEnum) {
+    final path = p.join(
+      'data',
+      'local_store',
+      'sql_store',
+      'tables',
+      '${model.name.snakeCase}.dart',
+    );
+    await _deleteFile(path);
   }
+}
+
+Future<void> _deleteFile(String path) async {
+  final file = File(path);
+  if (file.existsSync()) await file.delete();
 }
